requirements.txt
Flask==2.2.5
flask-cors==3.0.10
numpy==1.26.0
tensorflow==2.18.0
pillow==9.2.0
pydicom==2.3.0
opencv-python-headless==4.6.0.66
matplotlib==3.6.2